Move usePatronData to avoid dependency cycle problems
