+++ src/core/utils/helpers/usePatronData.ts
@@ -1,0 +1,9 @@
+import { useGetPatronInformationByPatronIdV2 } from "../../fbs/fbs";
+import { isAnonymous } from "./user";
+
+export const usePatronData = () =>
+  useGetPatronInformationByPatronIdV2({
+    enabled: !isAnonymous()
+  });
+
+export default {};
