+++ src/core/utils/withIsPatronBlockedHoc.tsx
@@ -7,8 +7,8 @@ import { setHasBeenVisible } from "../blockedModal.slice";
 import { RootState, useSelector } from "../store";
 import BlockedTypes from "./types/BlockedTypes";
 import { redirectTo } from "./helpers/url";
-import { usePatronData } from "./helpers/user";
 import { getModalIds } from "./helpers/modal-helpers";
+import { usePatronData } from "./helpers/usePatronData";
 
 export interface PatronProps {
   patron: AuthenticatedPatronV6 | null | undefined;
